Interface to set source spectrum now working
Spectrum of the light source can now be set using a spectral density
graph. May convert this to a plot of intensity rather than relative
proportions, for conceptual simplicity.
Need to embed within a wider div to make it easier to select ‘0’
brightness for any colour.
Also, must have a button for ‘uniform’ brightness, and somehow tie the
global brightness control to the spectral output graph.