Fixed mouse behaviour in spectral selector
Dragging outside the bounds of the spectrum now provides a reliable way
to set components of the spectrum to zero.

Things to do next:
1/do something similar at the top of the spectral graph to reliably set
100% intensity.
2/do something similar at the sides of the graph to allow those pesky
edge colours to be set accurately.
3/find out why the spectrum of the photons is always a little  higher in
energy than intended (getting magenta photons and it's very hard to get
red ones...)
4/increase the margins on the spectral graph region and add some axes
5/make the brightness in the graph absolute, rather than relative, and
tie the values to the brightness slider